# game-social-network-app/src/main/java/be/intecbrussel/gamesocialnetworkapp/responses/AuthenticationResponse.java
package be.intecbrussel.gamesocialnetworkapp.responses;

import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class AuthenticationResponse {

    private String token;
}