# game-social-network-ui/src/app/services/models/post-response.ts
/* tslint:disable */
/* eslint-disable */
export interface PostResponse {
  archived?: boolean;
  author?: string;
  content?: string;
  id?: number;
  image?: Array<string>;
  resume?: string;
  shareable?: boolean;
  title?: string;
}